Separate Draft Edit
My last attempt to separate draft edit. Everything seems to work fine.
